+++ core/templates/dev/head/components/forms/custom-forms-directives/apply-validation.directive.spec.ts
@@ -1,4 +1,4 @@
-// Copyright 2014 The Oppia Authors. All Rights Reserved.
+// Copyright 2019 The Oppia Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,237 +13,12 @@
 // limitations under the License.
 
 /**
- * @fileoverview Tests for the form builders.
+ * @fileoverview Tests for Directive for applying validation.
  */
 
 require(
-  'components/forms/custom-forms-directives/apply-validation.directive.ts');
-require('filters/convert-html-to-unicode.filter.ts');
-require('filters/convert-unicode-to-html.filter.ts');
-require('filters/convert-unicode-with-params-to-html.filter.ts');
-require(
-  'components/forms/custom-forms-directives/require-is-float.directive.ts');
-require('components/forms/validators/is-at-least.filter.ts');
-require('components/forms/validators/is-at-most.filter.ts');
-require('components/forms/validators/is-float.filter.ts');
-require('components/forms/validators/is-integer.filter.ts');
-require('components/forms/validators/is-nonempty.filter.ts');
-
-describe('HTML to text', function() {
-  beforeEach(angular.mock.module('oppia'));
-
-  var htmlUnicodeHtmlPairings = [
-    ['abc', 'abc', 'abc'],
-    ['&lt;a&copy;&deg;', '<a©°', '&lt;a&#169;&#176;'],
-    ['<b>a</b>', 'a', 'a'],
-    ['<br>a', 'a', 'a'],
-    ['<br/>a', 'a', 'a'],
-    ['<br></br>a', 'a', 'a'],
-    ['abc  a', 'abc  a', 'abc  a']
-  ];
-
-  it('should convert HTML to and from raw text correctly', angular.mock.inject(
-    function($filter) {
-      htmlUnicodeHtmlPairings.forEach(function(pairing) {
-        expect($filter('convertHtmlToUnicode')(pairing[0])).toEqual(pairing[1]);
-        expect($filter('convertUnicodeToHtml')(pairing[1])).toEqual(pairing[2]);
-      });
-    }
-  ));
-
-  var invalidUnicodeStrings = [
-    '{}',
-    '}}abc{{',
-    '\\{{a}}',
-    '{{a\\}}',
-    '{{a}\\}'
-  ];
-
-  it('should detect invalid unicode strings', angular.mock.inject(
-    function($filter) {
-      invalidUnicodeStrings.forEach(function(s) {
-        var fn = function() {
-          return $filter('convertUnicodeWithParamsToHtml')(s);
-        };
-        expect(fn).toThrow();
-      });
-    }));
-
-  var validUnicodeStrings = [
-    '{{}}',
-    '{{abc}}',
-    '\\\\{{abc}}',
-    '\\{{{abc}}'
-  ];
-
-  it('should detect valid unicode strings', angular.mock.inject(
-    function($filter) {
-      var results = [
-        '<oppia-parameter></oppia-parameter>',
-        '<oppia-parameter>abc</oppia-parameter>',
-        '\\<oppia-parameter>abc</oppia-parameter>',
-        '{<oppia-parameter>abc</oppia-parameter>',
-      ];
-      validUnicodeStrings.forEach(function(s, i) {
-        var fn = (function() {
-          return $filter('convertUnicodeWithParamsToHtml')(s);
-        })();
-        expect(fn).toEqual(results[i]);
-      });
-    }));
-});
-
-describe('Normalizer tests', function() {
-  var filterNames = [
-    'isFloat',
-    'isAtLeast',
-    'isAtMost',
-    'isNonempty',
-    'isInteger'
-  ];
-
-  beforeEach(angular.mock.module('oppia'));
-
-  it('should have the relevant filters', angular.mock.inject(function($filter) {
-    angular.forEach(filterNames, function(filterName) {
-      expect($filter(filterName)).not.toEqual(null);
-    });
-  }));
-
-  it('should validate floats correctly', angular.mock.inject(function($filter) {
-    var filter = $filter('isFloat');
-    expect(filter('1.23')).toEqual(1.23);
-    expect(filter('-1.23')).toEqual(-1.23);
-    expect(filter('0')).toEqual(0);
-    expect(filter('-1')).toEqual(-1);
-    expect(filter('-1.0')).toEqual(-1);
-    expect(filter('1,5')).toEqual(1.5);
-    expect(filter('1%')).toEqual(0.01);
-    expect(filter('1.5%')).toEqual(0.015);
-    expect(filter('-5%')).toEqual(-0.05);
-    expect(filter('.35')).toEqual(0.35);
-    expect(filter(',3')).toEqual(0.3);
-    expect(filter('.3%')).toEqual(0.003);
-    expect(filter('2,5%')).toEqual(0.025);
-    expect(filter('3.2% ')).toEqual(0.032);
-    expect(filter(' 3.2% ')).toEqual(0.032);
-    expect(filter('0.')).toEqual(0);
-
-    expect(filter('3%%')).toBeUndefined();
-    expect(filter('-')).toBeUndefined();
-    expect(filter('.')).toBeUndefined();
-    expect(filter(',')).toBeUndefined();
-    expect(filter('5%,')).toBeUndefined();
-    expect(filter('')).toBeUndefined();
-    expect(filter('1.23a')).toBeUndefined();
-    expect(filter('abc')).toBeUndefined();
-    expect(filter('2+3')).toBeUndefined();
-    expect(filter('--1.23')).toBeUndefined();
-    expect(filter('=1.23')).toBeUndefined();
-    expect(filter(undefined)).toBeUndefined();
-  }));
-
-  it('should impose minimum bounds', angular.mock.inject(function($filter) {
-    var filter = $filter('isAtLeast');
-    var args = {
-      minValue: -2.0
-    };
-    expect(filter(1.23, args)).toBe(true);
-    expect(filter(-1.23, args)).toBe(true);
-    expect(filter(-1.99, args)).toBe(true);
-    expect(filter(-2, args)).toBe(true);
-    expect(filter(-2.01, args)).toBe(false);
-    expect(filter(-3, args)).toBe(false);
-  }));
-
-  it('should impose maximum bounds', angular.mock.inject(function($filter) {
-    var filter = $filter('isAtMost');
-    var args = {
-      maxValue: -2.0
-    };
-    expect(filter(-2, args)).toBe(true);
-    expect(filter(-2.01, args)).toBe(true);
-    expect(filter(-3, args)).toBe(true);
-    expect(filter(1.23, args)).toBe(false);
-    expect(filter(-1.23, args)).toBe(false);
-    expect(filter(-1.99, args)).toBe(false);
-  }));
-
-  it('should validate non-emptiness', angular.mock.inject(function($filter) {
-    var filter = $filter('isNonempty');
-    expect(filter('a')).toBe(true);
-    expect(filter('')).toBe(false);
-  }));
-
-  it('should validate integers', angular.mock.inject(function($filter) {
-    var filter = $filter('isInteger');
-    expect(filter('3')).toBe(true);
-    expect(filter('-3')).toBe(true);
-    expect(filter('3.0')).toBe(true);
-    expect(filter('3.5')).toBe(false);
-  }));
-});
-
-describe('Testing requireIsFloat directive', function() {
-  var $compile, scope, testInput;
-
-  beforeEach(angular.mock.module('oppia'));
-
-  beforeEach(angular.mock.inject(function($compile, $rootScope) {
-    scope = $rootScope.$new();
-    var element = '<form name="testForm">' +
-      '<input name="floatValue" type="number" ng-model="localValue" ' +
-      'require-is-float apply-validation>' +
-      '</form>';
-    scope.validators = function() {
-      return [{
-        id: 'isFloat'
-      }];
-    };
-    $compile(element)(scope);
-    testInput = scope.testForm.floatValue;
-  }));
-
-  it('should validate if value is a float', function() {
-    testInput.$setViewValue('2');
-    scope.$digest();
-    expect(testInput.$valid).toEqual(true);
-
-    testInput.$setViewValue('2.0');
-    scope.$digest();
-    expect(testInput.$valid).toEqual(true);
-
-    testInput.$setViewValue('3.5');
-    scope.$digest();
-    expect(testInput.$valid).toEqual(true);
-
-    testInput.$setViewValue('-3.5');
-    scope.$digest();
-    expect(testInput.$valid).toEqual(true);
-  });
-
-  it('should invalidate if value is not a float', function() {
-    testInput.$setViewValue('-abc');
-    scope.$digest();
-    expect(testInput.$valid).toEqual(false);
-
-    testInput.$setViewValue('3..5');
-    scope.$digest();
-    expect(testInput.$valid).toEqual(false);
-
-    testInput.$setViewValue('-2.abc');
-    scope.$digest();
-    expect(testInput.$valid).toEqual(false);
-
-    testInput.$setViewValue('0.3.5');
-    scope.$digest();
-    expect(testInput.$valid).toEqual(false);
-
-    testInput.$setViewValue(undefined);
-    scope.$digest();
-    expect(testInput.$valid).toEqual(false);
-  });
-});
+  'components/forms/custom-forms-directives/' +
+  'apply-validation.directive.ts');
 
 describe('Testing apply-validation directive', function() {
   var $compile, element, scope, testInput;
