+++ core/templates/dev/head/components/forms/validators/is-integer.filter.spec.ts
@@ -1,0 +1,37 @@
+// Copyright 2019 The Oppia Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS-IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+/**
+ * @fileoverview Tests for Validator to check if input is integer.
+ */
+
+require('components/forms/validators/is-integer.filter.ts');
+
+describe('Normalizer tests', function() {
+  var filterName = 'isInteger';
+
+  beforeEach(angular.mock.module('oppia'));
+
+  it('should have the relevant filters', angular.mock.inject(function($filter) {
+    expect($filter(filterName)).not.toEqual(null);
+  }));
+
+  it('should validate integers', angular.mock.inject(function($filter) {
+    var filter = $filter('isInteger');
+    expect(filter('3')).toBe(true);
+    expect(filter('-3')).toBe(true);
+    expect(filter('3.0')).toBe(true);
+    expect(filter('3.5')).toBe(false);
+  }));
+});
